pyproject.toml
[project]
name = "qween3-5"
version = "0.1.0"
description = "Qwen3.5 2B fine-tuning for Verilog (communication interfaces + file ops)"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "mlx-lm>=0.21.0",
    "mlx>=0.25.0",
    "datasets>=4.7.0",
    "huggingface-hub>=1.7.1",
    "transformers>=4.47.0",
    "tokenizers>=0.21.0",
    "sentencepiece>=0.2.0",
    "numpy>=1.26.0",
    "tqdm>=4.66.0",
    "rich>=13.7.0",
    "gitpython>=3.1.40",
    "pyyaml>=6.0.1",
    "jinja2>=3.1.3",
]

[project.scripts]
clone-data   = "scripts.clone_data:main"
prepare      = "scripts.prepare_dataset:main"
finetune     = "scripts.finetune:main"
benchmark    = "scripts.benchmark:main"
export-gguf  = "scripts.export_gguf:main"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["scripts"]
